Move enactment code to it's own file
